+++ QtSchoolCanteen/users.cpp
@@ -2,24 +2,24 @@
 
 // user base class
 
-user::user(): name(""), password(""), credit(0){}
+user::user(): name(""), password(""), credit(0.){}
 
-user::user(QString name, QString password, int credit):
+user::user(QString name, QString password, double credit):
 	name(name), password(password), credit(credit){}
 
 QString user::getName() const { return name; }
 
 QString user::getPassword() const { return password; }
 
-int user::getCredit() const { return credit; }
+double user::getCredit() const { return credit; }
 
-void user::addCredit(int amount) { credit += amount; }
+void user::addCredit(double amount) { credit += amount; }
 
 // employee derived class
 
 employee::employee(): user(), department(""){}
 
-employee::employee(QString name, QString password, int credit, QString department):
+employee::employee(QString name, QString password, double credit, QString department):
 	user(name, password, credit), department(department){}
 
 QString employee::getDepartment() const { return department; }
@@ -29,7 +29,7 @@ QString employee::getDepartment() const { return department; }
 
 student::student(): user(), field(""), discount(0){}
 
-student::student(QString name, QString password, int credit, QString field, int discount):
+student::student(QString name, QString password, double credit, QString field, int discount):
 	user(name, password, credit), field(field), discount(discount){}
 
 QString student::getField() const { return field; }
@@ -41,7 +41,7 @@ int student::getDiscount() const { return discount; }
 
 staff::staff(): user(), position(none){}
 
-staff::staff(QString name, QString password, int credit, int position):
+staff::staff(QString name, QString password, double credit, int position):
 	user(name, password, credit), position(position){}
 
 int staff::getPosition() const { return position; }
